pyproject.toml
[project]
name = "template-website"
version = "0.1.0"
description = "A templated website to scaffold from."
requires-python = ">=3.13"
dependencies = [
    "httpx-oauth>=0.16.1",
    "infisicalsdk>=1.0.9",
    "jinja2>=3.1.6",
    "litestar[cryptography,opentelemetry]>=2.16.0",
    "opentelemetry-api>=1.38.0",
    "opentelemetry-exporter-otlp-proto-http>=1.38.0",
    "opentelemetry-sdk>=1.38.0",
    "piccolo-admin>=1.11.0",
    "piccolo-api[authenticator,pynacl]>=1.6.0",
    "piccolo[all]>=1.27.1",
    "python-dotenv>=1.1.1",
    "saq[hiredis,web]>=0.25.1",
    "skelmis-commons>=1.6.1",
    "uvicorn>=0.35.0",
    "uvloop>=0.22.1",
    "watchfiles>=1.1.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "freezegun>=1.5.2",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "ruff>=0.12.2",
]
